(maint) Enable hash passed as type_expr to TypeDefinitionInstantiator
An PObjectType, PTypeSetType, or PTypeAliasType can all be created
using either a literal hash expression or a common ruby hash. The
TypeDefinitionInstantiator#create_named_type assumed that the argument
type was the former. This commit makes it possible to also pass the
hash (the change is needed to accommodate subsequent additions).