(PUP-5982) Acceptance: add host selection to mk_tmp_environment
In rare cases, we need to specify a host to create a tmp environment
upon. This might be an agent, where one needs to use a full environment
for non-master duties outside of agent that has been pluginsync'd.
In the test for this ticket we use this to create an environment on the
agents to test puppet parser validate against a default manifest.
[skip ci]