(PUP-5645) Preserve HOME env variable in gem provider
Prior to this patch, the HOME environment variable was stripped
before the gem command was executed by the provider.  Since the
gem command uses the HOME environment variable to locate the
appropriate gemrc file to load, this was causing it to look at
/.gemrc.  Since Puppet runs as root, it is far more reasonable for
a customer to expect to find the gemrc file in /root/.gemrc.  This
patch maintains the previously existing HOME environment variable,
which forces gem to look in /root/.gemrc during puppet runs.