(maint) Use singleton pattern for TypeParser
A new TypeParser was created in lots of places in the code while other
places, like the evaluator, used a locally cached instance. Since using
a cached instance is quite safe, it's better to maintain such an
instance as a singleton in the class itself. This commit adds this
singleton and replaces all calls to TypeParser.new.