(PUP-5909) Add a ObjectType description to all PAnyType classes
All complex objects that can participate in serialization needs an
ObjectType that describes the object. This commit adds such a type
to all PAnyType classes. The type is not complete (it does not describe
the implementation in full) but it contains the attribute definitions
that holds the information needed during serialization.