(PUP-5909) Add ObjectType generation from RGen::Ecore classes
This commit adds an RGen::TypeGenerator that is capable of generating
PObjectType instances from the ecore classifier found in each RGen
specific class. The generator is then used to generate an object type
for each of the ecore specific classes found in the Puppet::Pops::Model
package and map them to the namespace 'Puppet::AST'.

A special registration is also performed where the Object type
`Puppet::AST::Locator` is bound to the implementation class
`Puppet::Pops::Parser::Locator::Locator19`.