(PUP-1498) Fix scheduled_task acceptance test on ruby 1.9.3
Ruby 1.9.3 has an ecoding bug in REXML, where it disregards the text encoding on a
string in preference for the XML header encoding value.  This causes an issue because
the XML file is encoded as UTF8.  In this commit we modify the XML Encoding header
returned by schtasks.exe to be UTF-8 not UTF-16 if the ruby version matches 1.9x

Additionally fixed a typo in acceptance test as it was calling facter on all agents
instead of the agent under test

[ci skip]