(maint) Remove dead version of the 'defined' function.
The `defined` function was migrated to 4.x a long time ago. The
implementation in the 3.x stub lingered. It is basically unreachable
code, and it is now removed.

This removal was triggered by a coming refactoring of certain calls into
scope - and fixing them in more than one place was not the right thing
to do.