(maint) Refactor use of known_resource_types
Scope:

Before this the helper methods delegating to known_resource_types used a
helper module instead of calling environment.known_resource_types
directly.
This "call for a call" is wasteful, and was used on a critival
code path. Now these calls are explicit.

Compiler:

This makes the roundabout ways in which known_resource_types was
accessed in the compiler direct (i.e. calling
environment.known_resource_types, instead of via a method included in a
helper module containing only this method).

The code is now easier to understand and it removes an extra call for
every such call - in some cases multiple such calls.

This also removes the ast_nodes? method as an alias for
known_resource_types.nodes? as it was only used internally in the
compiler. It could just call that method directly. Subsequently testing
if mocking of ast_nodes? works was removed because that was just silly.

When this work was done, it was noted that the
integration/parser/compiler_spec.rb was a modified copy of the
unit/parser/compiler_spec.rb. There was a large overlap between the two
in terms of tested functionality. Therefore, this commit also removes
all duplicated unit tests from the compiler integration tests.