(maint) Remove sort of Struct members
This commit removes the existing sort of the Struct members since there
is no reason to sort them. The order is irrelevant when assignability is
computed. It is however relevant to retain the order when a Struct is
used to describe named parameters of a function and the unnamed
parameter tuple is inferred from that Struct.