(PUP-6445) Pin win32 gems to explicit versions
The Win32-Service gem was updated to 0.8.9 however the new gem pulled in a
dependant gem called ffi-win32-extensions which has a name collision on the
FFI:Pointer.read_wide_string method.  This causes Puppet to fail in mutiple
locations.  This commit pins the Win32-Service gem to a prior version without
the dependency.

Additionally, all other win32-xxx gems are pinned to explicit versions so that
Puppet may not experience a similar issue in the future.