(PUP-6257) support --allow-authorization-extensions
This commit updates `puppet cert sign` to accept
--allow-authorization-extensions, which permits the signing of CSRs with
authorization extensions.

This closes the loop for utilizing authorization extensions in PE as we
now have a way of signing auth extensions in a trusted way.

This commit comes with a minor refactoring of `ca.sign` and
`ca.check_internal_signing_policies` to take an options hash.