(PUP-6398) Isolate module_working_dir in PMT tests
Executing parallel specs on Windows is not reliable in part, because
the module tool shares file system state across specs.  Previously,
executing `rake parallel:spec` on Windows would fail, even when passing
on other platforms.

This commit is one of several that will allow specs to be executed in
parallel by:

* Fixing colliding directory read/writes in installer_spec and
  upgrader_spec by stubbing Puppet.settings[:module_working_dir]
  to a new directory before each test in installer_spec and in
  upgrader_spec. This issue only appears when running these tests
  concurrently

  unpacker_spec was changed to follow a similar pattern due to
  discussion about Puppet settings maybe becoming immutable after
  first load in the future.