(PUP-6398) work around nil $CHILD_STATUS on Windows
On Windows, $CHILD_STATUS is nil while the tests in
spec/unit/provider/service/*.rb. and pkg_spec are running.
This results in a "Can't modify frozen object"
runtime error.  However, the tests are able to pass
when run serially with the tests in capability_spec, for
example. The reason being, some tests in capability_spec
make a compile_to_catalog() call which, as a side
effect, generates a pid.  $CHILD_STATUS is able to
latch on to this pid, resolving the issue.

The bug shows up when the tests are run in isolation or in
parallel (on occasion, when the tests are grouped in
such a way and executed in such an order that
$CHILD_STATUS is not properly set by another test before
these tests run).

This commit makes the following changes:

* Replaces $CHILD_STATUS.exitstatus calls with stubbed / mocked
    Puppet::Util::Execution.exitstatus calls in execution_spec
    and pacman_spec

* In the other specs, it generates a pid for $CHILD_STATUS by
    spinning up a process.

    The previous strategy does not work for these files because
    they call into $CHILD_STATUS in lib/provider files, not just
    in the specs. And, Puppet::Util::Execution.exitstatus is a
    private class method, so that it  is not possible to outright
    replace these $CHILD_STATUS.exitstatus calls.  This may be
    resolved with some refactoring in the future.

    Also, require 'English' does not set $CHILD_STATUS on
    Windows as it does on other platforms.  After requiring,
    $CHILD_STATUS (and $?) is still nil.