(PUP-6440) Support sensitive data tagging
This commit adds a `sensitive` attribute to properties to indicate that
the `should` and `is` values of a given property are sensitive and
should be specially handled. This commit also adds a `redacted` field to
events to indicate that the data contained in this event was sensitive
and has been redacted to prevent information leakage.