(maint) Add a PuppetStack that adds stack entries to the backtrace
This adds a mechanism which can be used to make an entry
on the ruby stack that indiciates that it is a puppet stack frame.
The frame is visible in the ruby backtrace and PuppetStack provides a
stacktrace that is filtered to only contain .pp entries.

This adds the mechanism and a test. The mechanism is not yet in use.