(maint) Reset Puppet version after version_spec tests
During work on PUP-6467, it was found that the version_spec
tests were modifying Puppet.version value globally, which was
affecting tests in integration/indirector/facts/facter_spec.rb.
Facter was incorrectly reporting a Puppet version of 1.2.3,
instead of 4.6.0 during the tests, causing failures.

This commit reverts Puppet.version to its original value after
each test in version_spec.  It also adds a context block and
new test to clarify the functionality of the version setter.