(PUP-6457) Always attempt remount for out of sync options
Prior to this commit, it was possible for puppet to thing that a
mounted device was in sync with it's resource definition when it was
not. This was because if a resource had already been mounted, and a
remount was attempted with bad options, the remount would fail even
though the entry in /etc/fstab would be updated. Puppet was only
checking the resource options against etc/fstab which meant that
after this point, it would believe the mounted object to be in sync
with the puppet resource describing it, even though it was not.

Now, puppet will check the resource's options against both /etc/fstab
and the output of the mount command (on linux operating systems). This
means that it can now detect if the correct object is not mounted and
will attempt a remount.