(PUP-6398) Allow settings_spec to run parallel
Executing parallel specs on Windows is not reliable in part, because
the Puppet type system must have been initialized prior to these
settings specs running, which may not be the case when run parallel.
Executing `rake parallel:spec` on Windows would fail, even when
passing on other platforms.

This commit is one of several that will allow specs to be executed in
parallel.

* With Puppet.features.microsoft_windows? stubbed to return false the
  first time a settings catalog is built, it will autorequire the
  user, file and group type class definitions.  When doing this, it
  triggers a URI building call through Puppet::Util.path_to_uri to
  specify a full path to the actual type implementation code from
  puppet/type/{file,user,group}.rb

  So when run in isolation, prior to this commit, 3 tests would fail
  as a result of this process because it would try to use invalid
  non-Windows path to load files.

  However, when this test is run after other tests that have loaded
  the user / group / file types, then there is no problem as the stub
  doesn't impact load behavior.

  This fix also addresses being able to run these specs parallel