(PUP-6468) Add config file for MCO test
This commit updates the QA-563_windows_exit_mcollective acceptance
test to create a configuration file for MCO. The puppet-agent installation
does not currently create a default configuration for MCollective, so this
step is required in order for MCO to start correctly.

The work for establishing a config file is captured in the JIRA ticket
MCO-727. The MCO config included in this commit can be removed when MCO-727
has been completed.