(PUP-6083) Deprecate use of UC/Resourceref as <ref> in Class[<ref>]
Before this it was possible to reference classes using a bare
word with initial UC. When evaluating that, the evaluator must
resolve the type reference (for example `Foo`) which may lead to
resolution of `Foo` as a resource type. If at the same time, references
to resource types are validated (to exist), the user will later get an
error. (For the time being, such references are left unresolved until
actually used as a resource reference which makes it harder to report
the source location).

This commit adds a deprecation warning (or error/off) under the control
of the strict flag.