(PUP-6434) Generalize Sensitive contained type
Parameterized types may contain a fair amount of information about the
data they describe; this information can constitute an information leak
when used with the Sensitive type. Given the following scenario:

    $o = String("hunter2")
    notice(type($o)) #=> String[7, 7]

In this instance, the type of an instantiated string leaks the length of
that string; if this string is a password an attacker might be able to
use this type information to reduce the attack space when brute forcing
this password.

This commit removes this source of leakage by generalizing contained
types upon instantiation of parameterized Sensitive types. This amounts
to type erasure which might prove problematic, as this might remove the
ability to tightly constrain sensitive values.

    class secrets_handler(Sensitive[Array[String[4, 8]]] $pwlist) {
        notice($pwlist)
    }

    class { "secrets_handler":
        pwlist => Sensitive(['hi', 'longlonglong']) # not an error!
    }

When dealing with sensitive information there's always a tradeoff;
providing more metadata about some unknown value makes it easier to
guess the value.