(PUP-5949) Add ObjectType declarations to Param and ResourceTypeImpl
This commit adds the ObjectType declarations to the Param and
ResourceTypeImpl classes in the Puppet::Pops::Resource module and
the call to assert that they are instantiated during boot. A test
is also added to verify that an instance of the
Puppet::Resource::ResourceType3 class can be created in Puppet code.