(PUP-5849) Differentiate '(' first on line from other '('
Before this, a '(' on a new line would still be taken as a continuation
of the expression on the preceding line if possible in the grammar.

In order to make it possible to make such '(' always be taken as the
start of a new expression this commit includes a processing rule for the
LPRAR token. If a '(' is only preceded by whitespace on the same line
then it is emitted as as WSLPAR token. 

Calls (both function call style f(), and method call style x.f() now
requires that the '(' is on the same line as the preceding non
whitespace character ('f' in the example) to be recognized as start of
arguments in a call.

Parameter lists, and regular parenthesized expressions are unaffected by
this change.