(PUP-6531) improve service provider detection on Ubuntu
Rather than relying on version strings to determine which init-system is
in use, use the fact that the presence of /run/systemd/system (on
Debian-based systems generally, but I only can confirm for Ubuntu)
implies systemd is suitable.

Additionally, do not attempt to to use upstart in that case. However, if
systemd is not in use and upstart is the init-system (as detected by the
presence of the appropriate binaries as currently), treat upstart as
suitable.

Thanks to Dmitry Teselkin for the suggested Facter addition of
:running_on_systemd in LP: #1570472.

Signed-off-by: Nishanth Aravamudan <nish.aravamudan@canonical.com>