(PUP-5949) Enable auto-loading of pp-defined resource types
This commits alters the Runtime3Loader so that it first checks if
a resource creation is present at the location:

<env_root>/.resource_types/<resource-name>.pp

If such a file is present, it will be loaded, verified, and evaluated.
The resulting resource type is stored in the loader and no further
check is made.

Names of built-in resources will bypass the check.