(maint) install cfpropertylist gem on all platforms
The cfpropertylist gem is needed to run unit tests against code that
parses plist files. We've gotten around this with lots of mocking and
stubbing in the past, but it's coming up to bite us in
https://github.com/puppetlabs/puppet/pull/5119

The core issue is that Windows is not the "ruby" platform - it's
"mingw". This adds a new special "all" platform to project_data.yaml,
where we can add gems which shouldn't be platform-constrained.

Probably what we really need is more complex logic around whether
platform-specific gems are needed only at runtime or also for tests,
and around which platform-specific gems are installable on a broader
array of platforms. This unblocks the above PR, though.