(maint) Use resource context for audit events
To avoid leaking information in transaction events we redact values
before passing them to the event; this way the event class itself will
only see redacted information. However this was problematic because the
`#audit_event` method was relying on the contents of a given event to
deter if audited values differ which doesn't work terribly well when
you've redacted both values.

This commit changes `#audit_event` to accept the resource application
context and use that for detecting audit changes without having to
expose information to event objects.