(PUP-6527) Ignore hiera_include calls from environment compiler
A common pattern for hiera_include is to put it at topscope, which
causes environment compilation to fail as the behavior of hiera_include
(or hiera in general) isn't well-defined for the environment compiler.

In order to allow users to immediately start using application
management and direct change without needing to modify how they do
classification, we now override hiera_include in the environment
compiler to just warn that it's being ignored.