(PUP-6439) Copy sensitive context from resources to types
This commit updates Puppet::Type to copy sensitive parameters from a
given resource upon instantiation.

Performing this copy is is somewhat complex due to how parameters behave
for a given type instance. Parameters are lazily generated depending on
if a value is set for that parameters, but some oddly behaving types
(looking at you, file type) have parameters and properties that
force the creation of other parameters and properties. So not only
do some types have to copy the sensitive context from one
parameter/property to another, but we may in fact to eagerly generate
a given param/property in order to set this context.

To allow individual types to manage these responsibilities, this commit
implements `Puppet::Type#set_sensitive_parameters` such that subclasses
can override it to perform type specific sensitive data handling. This
allows a given type to intelligently copy the sensitive context from one
param to another.

In the future we should change how types behave to remove the need for
this juggling act, but in the short term this approach gives us a
workable (if imperfect) implementation and a route forward for unusual
types.