(PUP-5948) Integrate generate
This changes the generator to make all output appear in one common
directory under <env>/.resource_types.
The logic to output relative to the input file is retained but currently
unused (and untested).

This completes the implementation of the integration testing of generate
and Pcore based Resource Type Implementation.
All currently supported functionality is tested. (Use of reflected type
information is not yet used and test for that is pending).

A number of small adjustments were made, primarily in the handling of
title patterns where the pcore information has a simpler model (as it
does not support lambda based title patterns). A transformation is
needed to the internal API as it also expects symbols instead of
strings.

The generate template for resource types was modified as the
implementation backing the Pcore described resource type implementations
does not support creating new instances via a hash and instead takes
arguments by position.