(PUP-6087) Error when a capability mapping is redefined.
Before this commit redefining a capability mapping added both mappings to
the capability list. When producing and consuming resources only the
first mapping would be used. Now trying to add the second mapping
results in a ParseError.

In addition we previously evaluated mappings twice. Once before
evaluating main and a second time after evaluating applications. The
capability mappings were not properly cleared after the first
evaluation. As a consequence capability mappings were added twice. It
seems better if we only only evaluate them once so that they are
available for producing resources at a consistent place in compilation.