(PUP-6572) Codify the `site:modules:$basemodulepath` pattern as default
For a very long time, CS has been recommending and educating the
`modulepath = site:modules:$basemodulepath` pattern via
`environment.conf`. It's now codified as a CS standard:

https://github.com/puppetlabs/cs-standards/blob/master/control_repo_contents.md#directory-structure-and-contents

It's always felt a bit weird to be explaining to people why the best
practice wasn't the Puppet default, so how about we make it the default?
This should not impact existing users unless they're currently using
`$environment/site` for other purposes, and I'd wager that percentage is
rather low.