(maint) Fix path confusion in pcore resource type loader
Before this commit, the pcore resource type loader was set up to
search in the parent directory of the environments 'lib' directory.
Even if that search sometimes succeeded, it is hardly correct.

This commit ensures that the FileType loader used for the pcore
resource types is setup with a path that doesn't include 'lib' and
then uses that path rather than using dirname to get the parent path.