(PUP-6182) Require `nodes` param for application declarations
Previously, this parameter was marked as optional in the type system,
but would cause a failure when processing the catalog and attempting to
iterate the nodes hash. Since this was always implicitly required and
there's no good reason for it not to always be specified, we now
explicitly require it so as to cause a clear, early failure.