(PUP-6182) Fail when non-existent components have nodes assigned
Previously, non-existent components in the node mapping of an
application would simply be ignored. Because this frequently happens
when the user incorrectly types the component name, it can be especially
confusing as the only complaint will be that no node was mapped to the
*correct* component. We now explicitly fail if there are non-existent
components, and we do so before checking whether there are unmapped
components.