(PUP-6620) Move pre-suite setup of CA certs on Windows
Prior to this commit, the Puppet presuite would install the certs
needed to connect to forge over PMT for all test modes. This was needed as the
Windows MSI did not include certs. That changed for :aio in RE-6111, in which
the cert bundle was included in the MSI, but this pre-suite item remained. This
commit moves it to the :git mode presuite only, because it is no longer needed
for :aio, and in PA-461 was found to be masking a misconfiguration in the MSI
  which resulted in a regression in functionality.

Beaker will search for the file as specified in the config options.rb in the
subdirector for the specified task, so we also update the impacted options.rb
files with the updated paths.