(PUP-6582) Modify how closure_scope is initialized and obtained
Before this, when a function was loaded it was given the global_scope
bound into context as its closure scope.
This was bad because the functions are cashed in the loaders, and the
loaders are cached in an environment when the environment timeout > 0.
This led to functions accessing the scope that was used when they were
first loaded which results in accessing the wrong/stale global variables
(such as $facts from the first compiled node).

This commit changes this so that loaded 4.x ruby functions and puppet
functions are initialized with a `nil` closure scope. At runtime the
`:global_scope` in Puppet context is looked up and used when a Closure
or Function has a `nil` `embeddning_scope` or `closure_scope`
(respectively).