(PUP-6582) Remove binding of loaders to AST model elements
Before this commit, 4x loaders were bound to AST model elements. That
was less than ideal since an AST model element has the same life cycle
as the environment. A new compile on the same environment would hence
find the loader created and attached to the adapter during the first
compile.

This commit retains the adapter but instead of attaching the actual
loader instance, it attaches the name of the loader. The loader is
then retrieved using `scope.compiler.loaders[loader_name]`