(PUP-5623) Add return type to Callable type (functions and lambdas)
This commit adds a return_type attribute to the Callable Puppet type
and the necessary syntax additions to support it, both in the Puppet
language parser and in the ruby function dispatchers. Tests are also
added to validate that the return type is used for validating the
returned value from lambdas, puppet functions, and ruby functions.