(PUP-6619) Add test for version file
This commit adds a new test to ensure the presence of the version file
as laid down by the AIO in accordance with the puppet agent path specification
after a recent set of changes to packaging of the AIO impacted the VERSION file
creation and a regression was discovered in associated logic. This is added as
a new test (as opposed to an additional step in 'ensure_puppet-agent_paths') to
take advantage of the benefits of `skip_test` since this test only applies to
packaged artifacts and not git/gem which don't create a VERSION file.

Signed-off-by: Moses Mendoza <mendoza.moses@gmail.com>