(PUP-6586) Check auth success codes for Windows logon
The password getter in the Windows user provider issues a logon attempt to
verify the password and determine if the value needs synchronized. Prior to
this commit, any failure to logon would result in a false return from
Puppet::Util::Windows::User.password_is? which would trigger a password reset.
As (somewhat) documented in
https://msdn.microsoft.com/en-us/library/windows/desktop/ms681385(v=vs.85).aspx,
there are several error codes that indicate logon failure for a different
reason than incorrect username/password. After some investigation, it was
determined that these codes could be used to indicate a successful
authentication even if logon failed. This commit modifies the method to
explicitly check for these error codes and return 'true' if encountered, to
prevent unnecessary password resets.