(PUP-6608) Fix problem with inline classes not getting evaluated
Before this commit, two names `resolved_type` and `fully_qualified_type`
where used to denote the same value in the method #create_resources of
class `Runtime3ResourceSupport`. Depending on code-path, only the former
variable was initialized when the resource type was a class. The
uninitialized variable was then used to test if the resource should be
subjected to class evaluation. This commit ensures that only one name
is used for the variable.