(PUP-6613) Fix undef values in reources created by create_resources
Before this create_resources had an implementation that filtered out all
undef attribute values. That was bad because if a resource parameter
does not have a default, and an undef value was given, it would fail
with "no value given".

Simply removing the filtering moved the problem to validation of
relationships. There a test was made if a value was nil, but not if it
was :undef - which is the case in the 3.x API under some conditions.

Lastly, when meta parameters for relationships are being filtered they
were only filtered for nil entries. Now they also filter out :undef
entries.