(PUP-6623) Define an unwrap function
This commit adds an unwrap function for sensitive objects; this serves
to expose the plaintext value in case the wrapped value needs to be
operated on/interpolated/etc. While this does remove the safety measure
that the Sensitive object provides it also allows more general use of
the data type.