(PUP-6628) Access @should directly when calculating corrective change
Currently, the corrective change logic uses the normal `should` and
`should=` accessors when messing with property values. This can cause
failures in property `validate` and `should=` if `unmunge` does not
put the data into a state that `munge` or `validate` is happy with.

By using a small amount of metaprogramming magic, we can access the
guts of the property value directly, bypassing those layers of logic
and creating a cleaner "transplant" of the value.