(PUP-6582) Refactor Closure to be base class
This refactors Closure so that the two behaviors (Static/named and
dynamic) are now separate base classes inheriting from Closure.
Only named/static closures can be bound by the loaders and this type of
closure always references the top scope indirectly. The Dynamic closure
holds a reference to a scope and must never be bound to something that
survives the end of a compilation.