(PUP-6627) Redact sensitive file diffs
This commit changes the file resource type diff printing such that when
the file origin is sensitive the diff is redacted.

In order to make redaction work correctly for both source and content
origins this pull request pushes the `sensitive` field from
Puppet::Property to Puppet::Parameter. Because of the silly nature of
the file type and how it juggles values around, the source parameter can
be unexpectedly used like a property and thus actually has semantics as
a sensitive values. To enable this, this commit pushes the sensitive
semantics up to Puppet::Parameter.