(PUP-6582) Cache loaders in environment
Before this, loaders where not cached in the environment. This lead to
busy work as functions and other definitions were read from disk,
parsed, and added to the loaders. Earlier this was required as the
loaded elements (functions in particular) held on to their
closure_scope. Earlier work on this ticket (PUP-6582) changes the
function reference to closure_scope such that it is indirect (via a
name) for the statically loaded functions. This made it possible to
cache the loaders in an environment.

The change sets the loaders in an environment, or reuses them when a
request is made to create loaders for a particular environment.

This work opens up for further simplifications (the loaders can be
obtained from the environment instead of from the Puppet Context). This
is however not yet changed as that would have an impact on a number of
test scenarios.